Introduce PyImport class to represent import statements
A new class `PyImport` is introduced to represent python's import
statement (consisting of `import`, `from`, `as`). With that, we can
have a more structured code for constructing and querying the database.

Besides, we now are able to specify import statements for a symbol
consisting of more than two segments (including dots); for example,
for a symbol `lib2.models.some_class.SomeClass`, we can trigger an
import statement like `from lib2.models import some_class` for the
input token `some_class.SomeClass`, in accordance with Google's python
style guide (use-import-for-packages-and-modules-only).